+++ CHANGELOG.md
@@ -1,3 +1,50 @@
+## [6.0.0-stage.2](https://github.com/aziontech/bundler/compare/v6.0.0-stage.1...v6.0.0-stage.2) (2025-08-15)
+
+
+### Bug Fixes
+
+* improve storage cleanup and move bucket globals to server initialization ([#506](https://github.com/aziontech/bundler/issues/506)) ([0158de6](https://github.com/aziontech/bundler/commit/0158de621d46797f0ea4f8e4428ff95ad3225f29))
+
+## [6.0.0-stage.1](https://github.com/aziontech/bundler/compare/v5.4.0-stage.4...v6.0.0-stage.1) (2025-08-15)
+
+
+### ⚠ BREAKING CHANGES
+
+* All azion.config files across all presets have been updated to the new API v4 schema. 
+Existing configurations are incompatible and must be rewritten to match the new format.
+Older files will not work with this release.
+
+* feat!: update all preset azion.config files for API v4 compatibility (#505) ([b7beae9](https://github.com/aziontech/bundler/commit/b7beae9c1498e20cc2896b791bd1a605aba474a8)), closes [#505](https://github.com/aziontech/bundler/issues/505)
+
+## [5.4.0-stage.4](https://github.com/aziontech/bundler/compare/v5.4.0-stage.3...v5.4.0-stage.4) (2025-08-14)
+
+
+### Bug Fixes
+
+* add onlyGenerateConfig option and improve storage error message ([#504](https://github.com/aziontech/bundler/issues/504)) ([0bab091](https://github.com/aziontech/bundler/commit/0bab09139892dfb77e5ed97b6fdd715fb8113af6))
+
+## [5.4.0-stage.3](https://github.com/aziontech/bundler/compare/v5.4.0-stage.2...v5.4.0-stage.3) (2025-08-13)
+
+
+### Features
+
+* improvements, command replace and flag only-generate-config to build command ([#503](https://github.com/aziontech/bundler/issues/503)) ([b964349](https://github.com/aziontech/bundler/commit/b964349f9fb46b8965faca34426220e844c06348))
+
+## [5.4.0-stage.2](https://github.com/aziontech/bundler/compare/v5.4.0-stage.1...v5.4.0-stage.2) (2025-08-13)
+
+
+### Features
+
+* API V4 ([#500](https://github.com/aziontech/bundler/issues/500)) ([41e7940](https://github.com/aziontech/bundler/commit/41e7940c6cd90b2b3798466700161499ef285204))
+
+## [5.4.0-stage.1](https://github.com/aziontech/bundler/compare/v5.3.0...v5.4.0-stage.1) (2025-08-05)
+
+
+### Features
+
+* add additional WEB APIs to local environment ([830d7e1](https://github.com/aziontech/bundler/commit/830d7e1d278139aac6fa3c78f6fb590301a6c6bc))
+* new web apis local dev ([#501](https://github.com/aziontech/bundler/issues/501)) ([b49a9a7](https://github.com/aziontech/bundler/commit/b49a9a7b59cbe30e9f0d1e08c4a6aaec91ec251e))
+
 ## [5.3.0](https://github.com/aziontech/bundler/compare/v5.2.0...v5.3.0) (2025-07-15)